Add: implementacion de services
Upd: metodos no utilizados

+++ src/main/java/com/marvic/springm6/repository/ILiquidacionRepository.java
@@ -3,5 +3,5 @@ package com.marvic.springm6.repository;
 import com.marvic.springm6.entity.Liquidacion;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ILiquidacionRepository extends JpaRepository<Liquidacion,Integer> {
+public interface ILiquidacionRepository extends JpaRepository<Liquidacion,Long> {
 }
