+++ src/main/java/com/marvic/springm6/service/ILiquidacionService.java
@@ -7,7 +7,7 @@ import java.util.List;
 public interface ILiquidacionService {
     List<Liquidacion> listarLiquidaciones();
     Liquidacion crearLiquidacion(Liquidacion liquidacion);
-    Liquidacion buscarLiquidacionPorId(int idLiquidacion);
+    Liquidacion buscarLiquidacionPorId(long idLiquidacion);
     Liquidacion actualizarLiquidacion(Liquidacion liquidacion);
-    void eliminarLiquidacionPorId(int idLiquidacion);
+    void eliminarLiquidacionPorId(long idLiquidacion);
 }
